fixed gcc warnings

+++ src/GUI/BookWindow.cpp
@@ -93,7 +93,7 @@ bool BookWindow::updateWindow(const sf::Time frameTime) {
 void BookWindow::setPage(int index) {
 	g_resourceManager->playSound(m_sound, SOUND_PATH);
 	auto const& pages = m_item.getDocumentPageBeans();
-	if (index < 0 || index > pages.size() - 1) return;
+	if (index < 0 || index > static_cast<int>(pages.size()) - 1) return;
 
 	m_currentPage = index;
 	auto const& page = pages.at(m_currentPage);
